+++ CHANGELOG.md
@@ -70,6 +70,7 @@ _Code:_
 - **Go**
 
   - Error `ErrOverflow` is now deprecated in favor of `ErrOutOfMemory`, new error types were added ([#711](https://github.com/cossacklabs/themis/pull/711)).
+  - `SecureMessage` now produces signed messages of optimal length ([#775](https://github.com/cossacklabs/themis/pull/775)).
 
 - **Objective-C**